requirements.txt
pandas>=2.0.0
pyupbit>=0.2.33
python-dotenv>=1.0.0
ta>=0.10.2
requests>=2.31.0
pydantic>=2.0.0
streamlit>=1.24.0
plotly>=5.13.0
fastapi>=0.68.0
uvicorn>=0.15.0